+++ components/callbacks/fx.go
@@ -31,7 +31,7 @@ import (
 )
 
 var Module = fx.Module(
-	"plugin.callbacks",
+	"component.callbacks",
 	fx.Provide(ConfigProvider),
 	fx.Provide(CallbackExecutorOptionsProvider),
 	fx.Invoke(RegisterTaskSerializers),
