+++ components/nexusoperations/completion.go
