+++ components/nexusoperations/workflow/commands_test.go
@@ -35,8 +35,8 @@ import (
 	historypb "go.temporal.io/api/history/v1"
 	persistencespb "go.temporal.io/server/api/persistence/v1"
 	"go.temporal.io/server/common/dynamicconfig"
-	"go.temporal.io/server/plugins/nexusoperations"
-	opsworkflow "go.temporal.io/server/plugins/nexusoperations/workflow"
+	"go.temporal.io/server/components/nexusoperations"
+	opsworkflow "go.temporal.io/server/components/nexusoperations/workflow"
 	"go.temporal.io/server/service/history/hsm"
 	"go.temporal.io/server/service/history/tests"
 	"go.temporal.io/server/service/history/workflow"
